# Generate correct symbols.o for sparc-unknown-none-elf