@nnethercote 2024-10-03
This seems fine to me but the discussion in #130172 is complex enough that I will defer to...

r? @workingjubilee