@workingjubilee 2024-10-04
I mulled this over for a bit, basically wrote a blog post worth of my thoughts on how silly this case was, then did some research and found something interesting: LLVM recognizes a lot more than just "v8" and "v9", but fortunately they encode v8 and v9ness.

https://github.com/llvm/llvm-project/blob/876f661dbeb42a76767edfb1f36214baacb27fd4/clang/lib/Basic/Targets/Sparc.cpp#L67-L104

And apparently they actually encode V8Plus-ness as a feature which they use when doing machine code generation:

https://github.com/llvm/llvm-project/blob/876f661dbeb42a76767edfb1f36214baacb27fd4/llvm/lib/Target/Sparc/MCTargetDesc/SparcTargetStreamer.cpp#L27-L28

Apparently passing this feature is handled in the Clang driver:

https://github.com/llvm/llvm-project/blob/876f661dbeb42a76767edfb1f36214baacb27fd4/clang/lib/Driver/ToolChains/Arch/Sparc.cpp#L39-L74

We can decide we don't want to recognize all these different CPUs, but we should make a decision here.