@workingjubilee 2024-10-04
Ah apparently that codegen attribute is actually fairly recent: https://github.com/llvm/llvm-project/pull/98713

So not in LLVM 19... nevermind.

Okay, doing some fiddling, I am missing a detail: LLVM has a fairly... idiosyncratic... way of doing option parsing. So they do recognize all the V8+ CPUs independently, it's not _entirely_ up to Clang. Whew!