@thejpster 2024-10-04
But if this makes Leon 3 bare metal work and 32-bit on 64-bit SPARC Linux work, that's ok for now right?