@workingjubilee 2024-10-04
Yes, I'm just saying that it should recognize the other "v9" CPUs as well, which are these:
https://github.com/llvm/llvm-project/blob/eaff3a743406ff1636e6328e1ba1bc66318d53cb/llvm/lib/Target/Sparc/Sparc.td#L148-L160

So, v9, ultrasparc, ultrasparc3, niagara, niagara2, niagara3, niagara4.