@thejpster 2024-10-04
I was unable to find a 32-bit SPARC gcc for Linux so sadly I can't test.