@taiki-e 2024-10-04
> I was unable to find a 32-bit SPARC gcc for Linux (Debian bookworm) so sadly I can't test.

FWIW, I used to be able to test using g++-multilib-sparc64 on Ubuntu/Debian and -m32 flag (https://github.com/taiki-e/setup-cross-toolchain-action/commit/ea9aca661e2713dbb7ee5c1c431cf2a8c31d4634), but since the nightly a while ago, the build has been broken with the "file in wrong format" error (https://github.com/taiki-e/setup-cross-toolchain-action/commit/8953e6c927f0185f159fa8bcb85a231893aee2e7, probably related to https://github.com/rust-lang/rust/commit/d1d21ede82614f2b575cb16bcbabe75183721740, considering when the error occurred).