@workingjubilee 2024-10-04
> The built in SPARC Linux target says `v9` so is this only an issue for anyone who overrides the target-cpu?

Correct.